+++ Assets/Weapons/Rocket Launcher/Scripts/RocketLauncherController.cs
@@ -8,7 +8,7 @@ namespace Weapons
     /// </summary>
     public class RocketLauncherController : WeaponController
     {
-        protected override void LaunchProjectile(float damage, GameObject owner)
+        protected override void LaunchProjectile(int damage, GameObject owner)
         {
             GameObject newProjectile_1 = Instantiate(projectile, transform.position, Quaternion.identity);
             GameObject newProjectile_2 = Instantiate(projectile, transform.position, Quaternion.identity);
